requirements.txt
Flask==1.1.2
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
python-dotenv==0.15.0
python-telegram-bot==13.2
